create release tag for HQ
This was missed when moving code from fabric in ce57162be4bc8e2bfa4378b6552ea140c3db4582
